+++ crates/bonsaidb-core/src/connection.rs
@@ -276,6 +276,15 @@ pub struct Collection<'a, Cn, Cl> {
     _phantom: PhantomData<Cl>, // allows for extension traits to be written for collections of specific types
 }
 
+impl<'a, Cn, Cl> Clone for Collection<'a, Cn, Cl> {
+    fn clone(&self) -> Self {
+        Self {
+            connection: self.connection,
+            _phantom: PhantomData,
+        }
+    }
+}
+
 impl<'a, Cn, Cl> Collection<'a, Cn, Cl>
 where
     Cn: Connection,
@@ -321,15 +330,14 @@ where
 
     /// Retrieves all documents matching `ids`. Documents that are not found
     /// are not returned, but no error will be generated.
-    pub fn list<R: Into<Range<u64>> + Send>(&self, ids: R) -> List<'_, Cn, Cl, R> {
-        List {
-            state: ListState::Pending(Some(ListBuilder {
-                collection: self,
-                range: ids,
-                sort: Sort::Ascending,
-                limit: None,
-            })),
-        }
+    pub async fn get_multiple(&self, ids: &[u64]) -> Result<Vec<OwnedDocument>, Error> {
+        self.connection.get_multiple::<Cl>(ids).await
+    }
+
+    /// Retrieves all documents matching `ids`. Documents that are not found
+    /// are not returned, but no error will be generated.
+    pub fn list<R: Into<Range<u64>>>(&'a self, ids: R) -> List<'a, Cn, Cl> {
+        List::new(PossiblyOwned::Borrowed(self), ids.into())
     }
 
     /// Removes a `Document` from the database.
@@ -341,29 +349,57 @@ where
     }
 }
 
-struct ListBuilder<'a, Cn, Cl, R> {
-    collection: &'a Collection<'a, Cn, Cl>,
-    range: R,
+pub(crate) struct ListBuilder<'a, Cn, Cl> {
+    collection: PossiblyOwned<'a, Collection<'a, Cn, Cl>>,
+    range: Range<u64>,
     sort: Sort,
     limit: Option<usize>,
 }
 
-enum ListState<'a, Cn, Cl, R> {
-    Pending(Option<ListBuilder<'a, Cn, Cl, R>>),
+pub(crate) enum PossiblyOwned<'a, Cl> {
+    Owned(Cl),
+    Borrowed(&'a Cl),
+}
+
+impl<'a, Cl> Deref for PossiblyOwned<'a, Cl> {
+    type Target = Cl;
+
+    fn deref(&self) -> &Self::Target {
+        match self {
+            PossiblyOwned::Owned(value) => value,
+            PossiblyOwned::Borrowed(value) => value,
+        }
+    }
+}
+
+pub(crate) enum ListState<'a, Cn, Cl> {
+    Pending(Option<ListBuilder<'a, Cn, Cl>>),
     Executing(BoxFuture<'a, Result<Vec<OwnedDocument>, Error>>),
 }
 
 /// Executes [`Connection::list()`] when awaited. Also offers methods to
 /// customize the options for the operation.
-pub struct List<'a, Cn, Cl, R> {
-    state: ListState<'a, Cn, Cl, R>,
+#[must_use]
+pub struct List<'a, Cn, Cl> {
+    state: ListState<'a, Cn, Cl>,
 }
 
-impl<'a, Cn, Cl, R> List<'a, Cn, Cl, R>
-where
-    R: Into<Range<u64>> + Send + 'a + Unpin,
-{
-    fn builder(&mut self) -> &mut ListBuilder<'a, Cn, Cl, R> {
+impl<'a, Cn, Cl> List<'a, Cn, Cl> {
+    pub(crate) const fn new(
+        collection: PossiblyOwned<'a, Collection<'a, Cn, Cl>>,
+        range: Range<u64>,
+    ) -> Self {
+        Self {
+            state: ListState::Pending(Some(ListBuilder {
+                collection,
+                range,
+                sort: Sort::Ascending,
+                limit: None,
+            })),
+        }
+    }
+
+    fn builder(&mut self) -> &mut ListBuilder<'a, Cn, Cl> {
         if let ListState::Pending(Some(builder)) = &mut self.state {
             builder
         } else {
@@ -371,13 +407,13 @@ where
         }
     }
 
-    /// Queries the view in ascending order.
+    /// Lists documents by id in ascending order.
     pub fn ascending(mut self) -> Self {
         self.builder().sort = Sort::Ascending;
         self
     }
 
-    /// Queries the view in descending order.
+    /// Lists documents by id in descending order.
     pub fn descending(mut self) -> Self {
         self.builder().sort = Sort::Descending;
         self
@@ -390,11 +426,10 @@ where
     }
 }
 
-impl<'a, Cn, Cl, R> Future for List<'a, Cn, Cl, R>
+impl<'a, Cn, Cl> Future for List<'a, Cn, Cl>
 where
     Cn: Connection,
-    Cl: schema::Collection,
-    R: Into<Range<u64>> + Send + 'a + Unpin,
+    Cl: schema::Collection + Unpin,
 {
     type Output = Result<Vec<OwnedDocument>, Error>;
 
@@ -415,7 +450,7 @@ where
                 let future = async move {
                     collection
                         .connection
-                        .list::<Cl, R>(range, sort, limit)
+                        .list::<Cl, _>(range, sort, limit)
                         .await
                 }
                 .boxed();
@@ -518,7 +553,7 @@ where
             .await
     }
 
-    /// Executes the query and retrieves the results with the associated [`CollectionDocument`s](crate::schema::CollectionDocument).
+    /// Executes the query and retrieves the results with the associated [`CollectionDocument`s](crate::document::CollectionDocument).
     pub async fn query_with_collection_docs(self) -> Result<Vec<MappedCollectionDocument<V>>, Error>
     where
         V::Collection: SerializedCollection,